refactor(service): extract systemd + Windows helpers from cli.py → os_/service.py
